added authentication

+++ Projects/QuizManager/src/controllers/quiz.ts
@@ -28,10 +28,15 @@ const createQuiz=async (req:Request, res:Response, next:NextFunction) =>{
 const getQuiz=async (req:Request, res:Response, next:NextFunction) =>{
     try {
         const quizId=req.params.quizId;
-        const quiz=await Quiz.findById(quizId,{name:1, questions_list:1, answers:1});
+        const quiz=await Quiz.findById(quizId,{name:1, questions_list:1, answers:1, created_by:1});
         if(!quiz){
             const err=new ProjectError("Quiz not found");
             err.statusCode=404;
+            throw err;
+        }
+        if(req.userId!==quiz.created_by.toString()){
+            const err=new ProjectError("You are not Authorized");
+            err.statusCode=403;
             throw err;
         }
         const resp:ReturnResponse={status:"success", message:"Quiz retrieved successfully", data:{quiz}};
@@ -51,6 +56,11 @@ const updateQuiz=async (req:Request, res:Response, next:NextFunction) =>{
             err.statusCode=404;
             throw err;
         }
+        if(req.userId!==quiz.created_by.toString()){
+            const err=new ProjectError("You are not Authorized");
+            err.statusCode=403;
+            throw err;
+        }
         quiz.name=req.body.name;
         quiz.questions_list=req.body.questions_list;
         quiz.answers=req.body.answers;
@@ -66,6 +76,17 @@ const updateQuiz=async (req:Request, res:Response, next:NextFunction) =>{
 const deleteQuiz=async (req:Request, res:Response, next:NextFunction) =>{
     try {
         const quizId=req.params.quizId;
+        const quiz=await Quiz.findById(quizId);
+        if(!quiz){
+            const err=new ProjectError("Quiz not found");
+            err.statusCode=404;
+            throw err;
+        }
+        if(req.userId!==quiz.created_by.toString()){
+            const err=new ProjectError("You are not Authorized");
+            err.statusCode=403;
+            throw err;
+        }
         await Quiz.deleteOne({_id:quizId});
         const resp:ReturnResponse={status:"success", message:"Quiz deleted successfully", data:{}};
         res.status(200).send(resp);
@@ -81,6 +102,11 @@ const publishQuiz=async (req:Request, res:Response, next:NextFunction) =>{
         if(!quiz){
             const err=new ProjectError("Quiz not found");
             err.statusCode=404;
+            throw err;
+        }
+        if(req.userId!==quiz.created_by.toString()){
+            const err=new ProjectError("You are not Authorized");
+            err.statusCode=403;
             throw err;
         }
         quiz.is_published=true;
